// sdk/src/types.ts
export type Utxo = {
  txid: Buffer;
  index: number;
  value: number;
};
export type Network = 'mainnet' | 'testnet' | 'regtest';
// ...
export type WalletUtxo = {
  /**
   * A reference to the unspent transaction output (UTXO) to be spent for this transaction. This
   * must be a pay-to-taproot (P2TR) output whose internal public key corresponds to the private key
   * provided.
   * @see https://github.com/bitcoin/bips/blob/master/bip-0341.mediawiki#user-content-Taproot_key_path_spending_signature_validation
   */
  utxo: Utxo;
  /**
   * A 32 byte private key that corresponds to the internal public key of the UTXO being spent and
   * can therefore be used to produce a valid signature for the transaction.
   */
  privkey: Uint8Array;
};
// ...
export type BitcoinTransactionNoChange = {
  /**
   * The serialized transaction in hexadecimal format.
   */
  txHex: string;
  /**
   * The hash of the transaction used to identify the transaction.
   * @see https://learnmeabitcoin.com/technical/transaction/input/txid/
   */
  txid: Buffer;
  /**
   * The index of the DID UTXO created in the transaction.
   */
  didUtxoIndex?: number;
  /**
   * The value of the DID UTXO created in the transaction.
   */
  didUtxoValue?: number;
};
/**
 * A bitcoin transaction that creates a change output.
 */
export type BitcoinTransactionWithChange = BitcoinTransactionNoChange & {
  /** The index of the change output in the transaction */
  changeIndex: number;
  /** The value of the change output in satoshis */
  changeValue: number;
  /** The address used for the change output */
  changeAddress: string;
};
/**
 * A bitcoin transaction.
 * @see https://developer.bitcoin.org/reference/transactions.html#raw-transaction-format
 */
export type BitcoinTransaction =
  | BitcoinTransactionNoChange
  | BitcoinTransactionWithChange;

/**
 * A type guard function that determines if a transaction has a change output.
 */
export function isBitcoinTransactionWithChange(
  tx: BitcoinTransactionNoChange,
): tx is BitcoinTransactionWithChange {
  return 'changeIndex' in tx;
}

export type CommitRevealTransactions = {
  commitTransaction: BitcoinTransaction;
  revealTransaction: BitcoinTransactionNoChange;
};